requirements.txt
python-dotenv
requests
openai 
termcolor